Add files via upload
The Flow puzzle is inspired by the popular Flow Free game, and was written in MMBasic for the Color Maximite 2 computer. As well as the game, there is also a level editor program, MakeFlowPuzzle, that lets you create your own puzzles to solve. The Puzzles directory contains  a number of puzzles in 3 difficulty levels.